worker/transcriber/requirements.txt
openai-whisper
kafka-python
boto3